feat: user entity & userController & userService & userRespository

+++ src/main/java/com/example/post/service/PostService.java
@@ -1,16 +1,17 @@
-package com.example.posts.service;
+package com.example.post.service;
 
 import java.util.List;
 
 import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import org.springframework.util.Assert;
 import org.springframework.web.server.ResponseStatusException;
 
-import com.example.posts.dto.PostForm;
-import com.example.posts.entity.Post;
-import com.example.posts.repository.PostRepository;
+import com.example.post.dto.PostForm;
+import com.example.post.entity.Post;
+import com.example.post.repository.PostRepository;
+import com.example.user.entity.User;
+import com.example.user.service.UserService;
 
 import lombok.RequiredArgsConstructor;
 
@@ -19,16 +20,8 @@ import lombok.RequiredArgsConstructor;
 @Service
 public class PostService {
 	private final PostRepository postRepository;
-
-	@Transactional(readOnly = true)
-	public List<Post> fetchAll() {
-		return postRepository.findAllByOrderByCreatedAtDesc();
-	}
-
-	public Post createPost(PostForm.Request request) {
-		Post post = Post.convertPostFormRequestToPostEntity(request);
-		return postRepository.save(post);
-	}
+	private final UserService userService;
+	// private final JWTGenerator jwtGenerator;
 
 	@Transactional(readOnly = true)
 	public Post loadPostById(Long id) {
@@ -37,27 +30,35 @@ public class PostService {
 				"Post with id{" + id + "} Not Found", new IllegalArgumentException()));
 	}
 
-	public Post updatePost(PostForm.Request request) {
-		Post post = loadPostById(request.getId());
-		checkValidation(request.getPassword(), post);
+	@Transactional(readOnly = true)
+	public List<Post> fetchAll() {
+		return postRepository.findAllByOrderByCreatedAtDesc();
+	}
+
+	public Post createPost(PostForm.Request request, String username) {
+		User user = userService.loadUserByUsername(username);
+		Post post = Post.convertPostFormRequestToPostEntity(request);
+		user.addPost(post);
+		return postRepository.save(post);
+	}
+
+	public Post updatePost(PostForm.Request request, Long id, String username) {
+		Post post = loadPostById(id);
+		if (post.checkWriter(username)){
+			throw new ResponseStatusException(HttpStatus.UNAUTHORIZED, "권한 없음");
+		}
 		post.updatePost(request);
 		return postRepository.save(post);
 	}
 
-	public void removePost(Long id, String password) {
+	public void removePost(Long id, String username) {
 		Post post = loadPostById(id);
-		checkValidation(password, post);
+		if (post.checkWriter(username)){
+			throw new ResponseStatusException(HttpStatus.UNAUTHORIZED, "권한 없음");
+		}
 		postRepository.deleteById(id);
 	}
 
-	private void checkValidation(String password, Post post) {
-		if (!checkPassword(password, post)) {
-			throw new ResponseStatusException(HttpStatus.UNAUTHORIZED, "Wrong Password");
-		}
-	}
 
-	private boolean checkPassword(String password, Post post) {
-		return post.match(password);
-	}
 
 }
